BFT-457: Fix committee with rota in test
