+++ node/tools/src/tests.rs
@@ -2,35 +2,20 @@ use crate::{store, AppConfig};
 use rand::{distributions::Distribution, Rng};
 use tempfile::TempDir;
 use zksync_concurrency::{ctx, sync};
-use zksync_consensus_roles::validator::{self, testonly::Setup, LeaderSelectionMode};
+use zksync_consensus_roles::validator::testonly::Setup;
 use zksync_consensus_storage::{testonly, PersistentBlockStore};
 use zksync_consensus_utils::EncodeDist;
 use zksync_protobuf::testonly::{test_encode_all_formats, FmtConv};
 
 impl Distribution<AppConfig> for EncodeDist {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> AppConfig {
-        let mut genesis: validator::GenesisRaw = rng.gen();
-        // In order for the genesis to be valid, the sticky leader needs to be in the validator committee.
-        if let LeaderSelectionMode::Sticky(_) = genesis.leader_selection {
-            let i = rng.gen_range(0..genesis.validators.len());
-            genesis.leader_selection =
-                LeaderSelectionMode::Sticky(genesis.validators.get(i).unwrap().key.clone());
-        } else if let LeaderSelectionMode::Rota(pks) = genesis.leader_selection {
-            let n = pks.len();
-            let i = rng.gen_range(0..genesis.validators.len());
-            let mut pks = Vec::new();
-            for _ in 0..n {
-                pks.push(genesis.validators.get(i).unwrap().key.clone());
-            }
-            genesis.leader_selection = LeaderSelectionMode::Rota(pks);
-        }
         AppConfig {
             server_addr: self.sample(rng),
             public_addr: self.sample(rng),
             debug_addr: self.sample(rng),
             metrics_server_addr: self.sample(rng),
 
-            genesis: genesis.with_hash(),
+            genesis: rng.gen(),
             max_payload_size: rng.gen(),
             validator_key: self.sample_opt(|| rng.gen()),
